requirements.txt
PyQt5==5.15.11
opencv-python==4.12.0.88
numpy==2.2.6
pyautogui==0.9.54
pynput==1.8.1
scikit-image==0.25.2
Pillow==12.0.0